# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "QuranAlsager API",
    "version": "1.0.0",
    "description": "API documentation for QuranAlsager"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/questions/{id}": {
      "get": {
        "tags": ["Questions"],
        "summary": "Get question by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Question ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Question found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          },
          "404": {
            "description": "Question not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "tags": ["Questions"],
        "summary": "Update question",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Question ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question_text": {
                    "type": "string",
                    "description": "The question text"
                  },
                  "option1": {
                    "type": "string",
                    "description": "First option"
                  },
                  "option2": {
                    "type": "string",
                    "description": "Second option"
                  },
                  "option3": {
                    "type": "string",
                    "description": "Third option"
                  },
                  "correctOption": {
                    "type": "string",
                    "description": "The correct option"
                  },
                  "timeLimitSec": {
                    "type": "number",
                    "description": "Time limit in seconds"
                  },
                  "categoryId": {
                    "type": "string",
                    "description": "ID of the category"
                  },
                  "ageGroupId": {
                    "type": "string",
                    "description": "ID of the age group"
                  },
                  "status": {
                    "type": "boolean",
                    "description": "Question status"
                  },
                  "app": {
                    "type": "boolean",
                    "description": "Flag for app availability"
                  },
                  "platform": {
                    "type": "boolean",
                    "description": "Flag for platform availability"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          },
          "404": {
            "description": "Question not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": ["Questions"],
        "summary": "Delete question (soft delete)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Question ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Question deleted successfully"
          },
          "404": {
            "description": "Question not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer token in the format 'Bearer {token}'"
      }
    },
    "schemas": {
...
      "Question": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Question ID"
          },
          "question_text": {
            "type": "string",
            "description": "The question text"
          },
          "option1": {
            "type": "string",
            "description": "First option"
          },
          "option2": {
            "type": "string",
            "description": "Second option"
          },
          "option3": {
            "type": "string",
            "description": "Third option"
          },
          "correctOption": {
            "type": "string",
            "description": "The correct option"
          },
          "timeLimitSec": {
            "type": "number",
            "description": "Time limit in seconds"
          },
          "status": {
            "type": "boolean",
            "description": "Question status"
          },
          "categoryId": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "ageGroupId": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "is_deleted": {
            "type": "boolean",
            "description": "Soft delete status"
          },
          "deleted_by": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              }
            }
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_by": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              }
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}